+++ hutool-core/src/main/java/cn/hutool/core/text/StrTemplate.java
@@ -1,4 +1,0 @@
-package cn.hutool.core.text;
-
-public class StrTemplate {
-}